move messaging services to folders
