+++ src/api/organization.js
@@ -41,6 +41,16 @@ export const schema = gql`
     unsetFeatures: [String]
   }
 
+  enum MessagingServiceType {
+    SMS
+  }
+
+  type MessagingService {
+    name: String!
+    type: MessagingServiceType
+    config: JSON
+  }
+
   input OrgSettingsInput {
     messageHandlers: [String]
     actionHandlers: [String]
@@ -76,6 +86,7 @@ export const schema = gql`
     twilioAccountSid: String
     twilioAuthToken: String
     twilioMessageServiceSid: String
+    messagingService: MessagingService
     fullyConfigured: Boolean
     emailEnabled: Boolean
     phoneInventoryEnabled: Boolean!
