+++ src/extensions/messaging_services/nexmo/index.js
@@ -1,8 +1,8 @@
 import Nexmo from "nexmo";
-import { getFormattedPhoneNumber } from "../../lib/phone-format";
-import { Message, PendingMessagePart } from "../../server/models";
-import { getLastMessage } from "./message-sending";
-import { log } from "../../lib";
+import { getFormattedPhoneNumber } from "../../../lib/phone-format";
+import { Message, PendingMessagePart } from "../../../server/models";
+import { getLastMessage } from "../message-sending";
+import { log } from "../../../lib";
 
 // NEXMO error_codes:
 // If status is a number, then it will be the number
